+++ appointment-service/src/main/java/co/unicauca/piedrazul/appointment/domain/model/UserCache.java
@@ -6,7 +6,7 @@ package co.unicauca.piedrazul.appointment.domain.model;
  */
 public class UserCache {
 
-    private int userId;
+    private long userId;
     private String fullName;
     private String role;
     private String state;
@@ -17,8 +17,8 @@ public class UserCache {
         return "ACTIVO".equalsIgnoreCase(state);
     }
 
-    public int getUserId() { return userId; }
-    public void setUserId(int userId) { this.userId = userId; }
+    public long getUserId() { return userId; }
+    public void setUserId(long userId) { this.userId = userId; }
     public String getFullName() { return fullName; }
     public void setFullName(String name)  { this.fullName = name; }
     public String getRole() { return role; }
